Adds multiprocessing support
In order to improve the speed of rendering the scenes this
modifies the rendering engine to run multiple processes in order
to render several rows at the same time.

The engine now returns an iterable which allows any consumer to
decide how to handle these returned rows and add any progress
bar functionality if desired.