Migrates slow components to Cython
Trying to get the timings down and the maths in the Vector
and Colour classes are called a lot so moving them to Cython.

By using this have more than halved the timings to render a scene,
the biggest win was migrating the Colour class.

Also a thanks to @philip238